Convert IEEE fetcher to new interface (#4162)
* Convert IEEE fetcher to new interface

* Move JSON parsing to respective fetcher

* Remove obsolete localization keys
