+++ src/main/java/org/jabref/logic/importer/WebFetchers.java
@@ -90,6 +90,7 @@ public class WebFetchers {
         list.add(new SpringerFetcher());
         list.add(new CrossRef());
         list.add(new DOAJFetcher(importFormatPreferences));
+        list.add(new IEEE(importFormatPreferences));
         list.sort(Comparator.comparing(WebFetcher::getName));
         return list;
     }
@@ -139,7 +140,7 @@ public class WebFetchers {
         fetchers.add(new SpringerLink());
         fetchers.add(new ACS());
         fetchers.add(new ArXiv(importFormatPreferences));
-        fetchers.add(new IEEE());
+        fetchers.add(new IEEE(importFormatPreferences));
         // Meta search
         fetchers.add(new GoogleScholar(importFormatPreferences));
         fetchers.add(new OpenAccessDoi());
