+++ src/main/resources/l10n/JabRef_en.properties
@@ -568,7 +568,6 @@ Importing=Importing
 
 Importing\ in\ unknown\ format=Importing in unknown format
 
-Include\ abstracts=Include abstracts
 Include\ entries=Include entries
 
 Include\ subgroups\:\ When\ selected,\ view\ entries\ contained\ in\ this\ group\ or\ its\ subgroups=Include subgroups: When selected, view entries contained in this group or its subgroups
@@ -709,9 +708,6 @@ no\ base-BibTeX-file\ specified=no base-BibTeX-file specified
 no\ library\ generated=no library generated
 
 No\ entries\ found.\ Please\ make\ sure\ you\ are\ using\ the\ correct\ import\ filter.=No entries found. Please make sure you are using the correct import filter.
-
-
-No\ entries\ found\ for\ the\ search\ string\ '%0'=No entries found for the search string '%0'
 
 No\ entries\ imported.=No entries imported.
 
@@ -1176,7 +1172,6 @@ XMP-metadata=XMP-metadata
 XMP-metadata\ found\ in\ PDF\:\ %0=XMP-metadata found in PDF: %0
 You\ must\ restart\ JabRef\ for\ this\ to\ come\ into\ effect.=You must restart JabRef for this to come into effect.
 You\ have\ changed\ the\ language\ setting.=You have changed the language setting.
-You\ have\ entered\ an\ invalid\ search\ '%0'.=You have entered an invalid search '%0'.
 
 You\ must\ restart\ JabRef\ for\ the\ new\ key\ bindings\ to\ work\ properly.=You must restart JabRef for the new key bindings to work properly.
 
@@ -1488,7 +1483,6 @@ Toggle\ print\ status=Toggle print status
 Update\ keywords=Update keywords
 Write\ values\ of\ special\ fields\ as\ separate\ fields\ to\ BibTeX=Write values of special fields as separate fields to BibTeX
 You\ have\ changed\ settings\ for\ special\ fields.=You have changed settings for special fields.
-%0\ entries\ found.\ To\ reduce\ server\ load,\ only\ %1\ will\ be\ downloaded.=%0 entries found. To reduce server load, only %1 will be downloaded.
 A\ string\ with\ that\ label\ already\ exists=A string with that label already exists
 Connection\ to\ OpenOffice/LibreOffice\ has\ been\ lost.\ Please\ make\ sure\ OpenOffice/LibreOffice\ is\ running,\ and\ try\ to\ reconnect.=Connection to OpenOffice/LibreOffice has been lost. Please make sure OpenOffice/LibreOffice is running, and try to reconnect.
 JabRef\ will\ send\ at\ least\ one\ request\ per\ entry\ to\ a\ publisher.=JabRef will send at least one request per entry to a publisher.
