+++ src/test/java/org/jabref/logic/importer/fetcher/SpringerFetcherTest.java
@@ -2,10 +2,12 @@ package org.jabref.logic.importer.fetcher;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.BibtexEntryTypes;
 
+import org.json.JSONObject;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -38,5 +40,28 @@ class SpringerFetcherTest {
 
         List<BibEntry> fetchedEntries = fetcher.performSearch("JabRef Social Barriers Steinmacher");
         assertEquals(Collections.singletonList(expected), fetchedEntries);
+    }
+
+    @Test
+    void testSpringerJSONToBibtex() {
+        String jsonString = "{\r\n" + "            \"identifier\":\"doi:10.1007/BF01201962\",\r\n"
+                + "            \"title\":\"Book reviews\",\r\n"
+                + "            \"publicationName\":\"World Journal of Microbiology & Biotechnology\",\r\n"
+                + "            \"issn\":\"1573-0972\",\r\n" + "            \"isbn\":\"\",\r\n"
+                + "            \"doi\":\"10.1007/BF01201962\",\r\n" + "            \"publisher\":\"Springer\",\r\n"
+                + "            \"publicationDate\":\"1992-09-01\",\r\n" + "            \"volume\":\"8\",\r\n"
+                + "            \"number\":\"5\",\r\n" + "            \"startingPage\":\"550\",\r\n"
+                + "            \"url\":\"http://dx.doi.org/10.1007/BF01201962\",\"copyright\":\"©1992 Rapid Communications of Oxford Ltd.\"\r\n"
+                + "        }";
+
+        JSONObject jsonObject = new JSONObject(jsonString);
+        BibEntry bibEntry = SpringerFetcher.parseSpringerJSONtoBibtex(jsonObject);
+        assertEquals(Optional.of("1992"), bibEntry.getField("year"));
+        assertEquals(Optional.of("5"), bibEntry.getField("number"));
+        assertEquals(Optional.of("#sep#"), bibEntry.getField("month"));
+        assertEquals(Optional.of("10.1007/BF01201962"), bibEntry.getField("doi"));
+        assertEquals(Optional.of("8"), bibEntry.getField("volume"));
+        assertEquals(Optional.of("Springer"), bibEntry.getField("publisher"));
+        assertEquals(Optional.of("1992-09-01"), bibEntry.getField("date"));
     }
 }
